refactor: consolidate path_utils — delete lib/path_utils.py, core/ is SSOT
Move path_matches_pattern, parent_path, unscope_internal_path from
lib/path_utils.py to core/path_utils.py. Delete lib/path_utils.py.
All callers redirect to core. Brick import boundary whitelist updated.
