+++ .pre-commit-hooks/check_brick_imports.py
@@ -37,11 +37,15 @@ BRICKS_RELATIVE_PATH = Path("src") / "nexus" / "bricks"
 FORBIDDEN_PATTERNS: list[tuple[re.Pattern[str], str]] = [
     # Direct core imports (excluding protocols and storage-pillar ABCs)
     (
-        re.compile(r"^\s*from\s+nexus\.core(?!\.protocols\b|\.cache_store\b|\.object_store\b)"),
+        re.compile(
+            r"^\s*from\s+nexus\.core(?!\.protocols\b|\.cache_store\b|\.object_store\b|\.path_utils\b)"
+        ),
         "nexus.core",
     ),
     (
-        re.compile(r"^\s*import\s+nexus\.core(?!\.protocols\b|\.cache_store\b|\.object_store\b)"),
+        re.compile(
+            r"^\s*import\s+nexus\.core(?!\.protocols\b|\.cache_store\b|\.object_store\b|\.path_utils\b)"
+        ),
         "nexus.core",
     ),
     # Direct services imports (protocols moved to contracts/protocols/)
