+++ src/nexus/bricks/search/search_service.py
@@ -2490,7 +2490,7 @@ class SearchService:
         Raises:
             InvalidPathError: If path is invalid
         """
-        from nexus.lib.path_utils import validate_path
+        from nexus.core.path_utils import validate_path
 
         return validate_path(path, allow_root=True)
